Load touchstone data to database

Load touchstone-relevant transformed data into the database. This should be called from the load code of a montagu-import, which must therefore be a custom load, not an automatic one. Call stoner::stone_load(transformed_data, con) with the same arguments provided to the montagu-import load function.

Usage

stone_load(
  transformed_data,
  con,
  allow_overwrite_scenario_description = FALSE,
  allow_overwrite_scenario_type = FALSE
)

Arguments

transformed_data

The transformed data from the transform stage. In a form which adheres to the DB schema.

con

Active connection to the DB which the data will be loaded to.

allow_overwrite_scenario_description

If TRUE, then stoner will overwrite rows in the scenario_description table with rows from the CSV file, (matching id, different description or disease), even if there are open or finished touchstones that use this description.

allow_overwrite_scenario_type

If TRUE, then stoner will overwrite rows in the scenario_type table with rows from the CSV file, (matching id, different name), even if there are open or finished touchstones that use this description.