Upload stochastic data to annex

Upload the reduced csvs produced by stoner_stochastic_process into the annex database.

stone_stochastic_upload(
  file,
  con,
  annex,
  modelling_group,
  disease,
  touchstone,
  is_cohort,
  is_under5,
  allow_new_database = FALSE,
  testing = FALSE
)

Arguments

file

A csv file generated by stone_stochastic_process

con

DBI connection to production (for argument validation)

annex

DBI connection to annex, to receive stochastic uploads.

modelling_group

The modelling group id

disease

The disease

touchstone

The touchstone (including version) for these estimates

is_cohort

Set this to TRUE if the csv file is cohort-oriented, or FALSE for calendar-year oriented.

is_under5

Set this to TRUE if the csv file only considers ages 0-4 inclusive; FALSE if all ages are included.

allow_new_database

Create the stochastic_file database if it does not exist. Should only be needed first time on a new database.

testing

For internal use only.